requirements.txt

pymongo == 3.7.1
Pillow == 5.2.0
setuptools == 40.0.0


